# openapi.yaml
openapi: 3.0.3
info:
  title: Retirement‐Matching API
  description: |
    REST endpoints for user auth (via magic‐link), retirees, career elements,
    reference data, startups, job postings, matching, conversations, and hire invites.
    Uses a JWT set in a cookie.
  version: 1.0.0
servers:
  - url: 'https://api.example.com'
    description: Production server
  - url: 'http://localhost:4000'
    description: Local development
security:
  - JwtCookieAuth: []
paths:
...
  /expertise-area:
    get:
      summary: List all expertise areas
      responses:
        '200':
          description: Array of expertise areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpertiseArea'
      security:
        - JwtCookieAuth: []
  '/expertise-area/{expertiseAreaId}':
    get:
      summary: Get an expertise area by ID
      parameters:
        - name: expertiseAreaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expertise area object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertiseArea'
        '404':
          description: Not found
      security:
        - JwtCookieAuth: []
...
  /startup:
    get:
      summary: Get all startups
      responses:
        '200':
          description: List of startups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Startup'
      security:
        - JwtCookieAuth: []
  '/startup/{startupId}':
    get:
      summary: Get a startup by ID
      parameters:
        - name: startupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Startup object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Startup'
        '404':
          description: Not found
      security:
        - JwtCookieAuth: []
    put:
      summary: Update a startup
      parameters:
        - name: startupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Fields to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartupUpdate'
        required: true
      responses:
        '200':
          description: Updated startup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Startup'
        '400':
          description: Validation error
        '404':
          description: Not found
      security:
        - JwtCookieAuth: []
    delete:
      summary: Delete a startup
      parameters:
        - name: startupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Not found
      security:
        - JwtCookieAuth: []
...
components:
  schemas:
...
    User:
      type: object
      required:
        - _id
        - email
        - nameLast
        - nameFirst
      properties:
        _id:
          type: string
          description: Unique identifier for any user (retiree or startup).
          format: uuid
        userType:
          x-stoplight:
            id: xn0mj8gqsyg6a
          enum:
            - Retiree
            - Startup
        email:
          type: string
          description: Must be unique; used for magic‐link login.
          format: email
        passwordHash:
          type: string
          description: '(Optional) Stored hash if you ever add password auth. For now, not used.'
          nullable: true
        nameLast:
          type: string
          description: Last name of the retiree/startup user.
        nameFirst:
          type: string
          description: First name of the retiree/startup user.
        profilePicture:
          type: string
          nullable: true
        createdAt:
          type: string
          description: Date and time when the user was created.
          format: date-time
        notifications:
          type: array
          x-stoplight:
            id: djl70mwpr4wm7
          items:
            $ref: '#/components/schemas/Notification'
...
    Startup:
      allOf:
        - $ref: '#/components/schemas/User'
        - required:
            - title
          type: object
          properties:
            title:
              type: string
              description: The official name of the startup.
            aboutUs:
              type: string
              description: A description of the startup (‐About us‐).
            contactPersonNameLast:
              type: string
              description: Last Name of the main contact person.
            contactPersonNameFirst:
              type: string
              description: First Name of the main contact person.
            contactPersonPicture:
              type: string
              description: URL to the contact person's picture.
              nullable: true
            industry:
              type: string
              description: 'Industry or sector (e.g., ‐FinTech‐, ‐HealthTech‐).'
            addressCity:
              type: string
              description: City.
              nullable: true
            addressCountry:
              type: string
              description: Country.
              nullable: true
            fundingStatus:
              type: string
              format: uuid
            fulltimeEmployeesNum:
              type: integer
              description: Number of full‐time employees.
            foundingYear:
              type: integer
              description: Year the startup was founded.
            revenuePerYearEUR:
              type: number
              description: Annual revenue in EUR.
              format: float
            imprintUrl:
              type: string
              description: Link to the legally required imprint page (if applicable).
              format: uri
            websiteUrl:
              type: string
              description: Public website URL.
              format: uri
            logoUrl:
              type: string
              description: URL to the startup’s logo.
              nullable: true
            stripeCustomerId:
              type: string
              description: Stripe Customer ID.
              nullable: true
            stripePriceId:
              type: string
              description: Stripe Price ID.
              nullable: true
            stripeSubscriptionExpiryDate:
              type: string
              description: When the Stripe subscription runs out.
              format: date
              nullable: true
            wiseUpSubscriptionTier:
              type: string
              description: Current WiseUp subscription tier.
              nullable: false
              enum:
                - free
                - silver
                - gold
              default: free
            monthlyConnectionBalance:
              type: number
              description: Balance of ‐connections‐/credits the startup has on a monthly basis. Gets reset every month.
              format: int32
            permanentConnectionBalance:
              type: number
              description: Balance of ‐connections‐/credits the startup has on a permanent basis. Does not reset. only available as extension to the gold tier.
              format: int32
            jobPostings:
              type: array
              description: List of job postings for this startup.
              items:
                type: string
                format: uuid
          description: '‐Startup‐ profile, extends User.'
    StartupUpdate:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          description: '‐Startup‐ profile, extends User.'
          properties:
            title:
              type: string
            aboutUs:
              type: string
            contactPersonNameLast:
              type: string
              description: Last Name of the main contact person.
            contactPersonNameFirst:
              type: string
              description: First Name of the main contact person.
            contactPersonPicture:
              type: string
              nullable: true
            industry:
              type: string
            addressCity:
              type: string
              description: City.
              nullable: true
            addressCountry:
              type: string
              description: Country.
              nullable: true
            fundingStatus:
              type: string
              format: uuid
            fulltimeEmployeesNum:
              type: integer
            foundingYear:
              type: integer
            revenuePerYearEUR:
              type: number
              format: float
            imprintUrl:
              type: string
              format: uri
            websiteUrl:
              type: string
              format: uri
            logoUrl:
              type: string
              nullable: true
            jobPostings:
              type: array
              description: List of job postings for this startup.
              items:
                type: string
                format: uuid
...
    ExpertiseArea:
      required:
        - _id
        - name
      type: object
      properties:
        _id:
          type: string
          description: UUID for this expertise area (e.g. ‐Data Science‐).
          format: uuid
        name:
          type: string
          description: Human‐readable name of the expertise area.
...
    Notification:
      title: Notification
      type: object
      required:
        - _id
        - title
        - message
        - actions
        - timestamp
      properties:
        _id:
          type: string
        title:
          type: string
          x-stoplight:
            id: d49j93kvvm3i3
        message:
          type: string
          x-stoplight:
            id: mxq5ow0lg3y3q
        actions:
          type: array
          x-stoplight:
            id: ali6qn5w8m0lc
          items:
            x-stoplight:
              id: n3n9c4nepvlrr
            type: object
            required:
              - label
              - url
            properties:
              label:
                type: string
                x-stoplight:
                  id: r5q0jvzmxjn3d
              url:
                type: string
                x-stoplight:
                  id: 6nb7iti38g0ke
        timestamp:
          type: string
          x-stoplight:
            id: kaes30f7rtl5r
  securitySchemes:
    JwtCookieAuth:
      type: apiKey
      name: jwt
      in: cookie